irepa/assets/flux/requirements.txt
## Requirements for Flux image generation scripts

replicate>=0.25.0
pyyaml>=6.0
pillow>=10.0.0